requirements.txt
Levenshtein==0.25.1
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.8.2
python-Levenshtein==0.25.1
pytz==2024.1
rapidfuzz==3.8.1
six==1.16.0
tzdata==2024.1
xlrd==2.0.1
